When user creation is blocked, we get a defined CurrentUser
RT-Authen-ExternalAuth can block user creation in the mailgate by
returning 0 from CanonicalizeUserInfo if the user doesn't exist in the
external system.  RT blindly assumes that
RT::Interface::Email::CreateUser will return a user, so it returns the
CurrentUser and 1 without checking to see if the creation was
successful.

This results in CommandByMail getting a defined but unloaded CurrentUser

Checking for the Id prevents later failures