Two pod burps going pod->text for the README
Stop quoting everything that is in C<> (we should fold this into
Module::Install::RTx and then we can clean up this Makefile.PL)

Restructure the BUGS section so that long long L<>s can't wrap, although
this adds a lot of whitespace.  This (or some similar variant) belongs
in Dist::Zill::MintingProfile::RTx.