Clarify header parsing features
The previous documentation implied that the extension looked
equally at headers and the email body for possible commands
to process. However, the extent of header parsing is scoped to
defining a specific header and then toggling whether you want the
email body to also be parsed or not when using that header. This
commit clarifies the functionality.