Switch from "use utf8" to explicit decode(), for clarity
Use of "use utf8" requires that the programmer constantly be aware
that string literals may be treated with character semantics, instead
of byte semantics, depending on lexical scope.  Make clear that the
return of ->Subject contains characters, not bytes, by leaving them as
bytes in the literal, and explicitly decoding them.

See 7548587b in RT.