Make pseudoheader whitespace matches non-greedy
The regex for matching pseudoheaders in the message body had three
places where it specified '\s*?', a non-greedy 0-or-more match on
whitespace.

In two of these cases, the one before the colon ':' and the one near the
end of the regex, due to the context around the whitespace a non-greedy
match MUST match the exact same substring as a greedy match would,
making them unnecessary and possibly causing additional backtracking.

In the third case, the one between the colon ':' and the second capture
group '(.*)', the non-greedy qualifier means that it will relinquish all
whitespace to be included in the second capture group.  This defeats the
purpose of having the '\s' at that position in the expression.